+++ haskell-src/Concordium/Genesis/Data/Base.hs
@@ -243,6 +243,7 @@ toChainParameters genesisAccounts GenesisChainParameters{..} = ChainParameters{.
     _cpConsensusParameters = gcpConsensusParameters
     _cpFinalizationCommitteeParameters = gcpFinalizationCommitteeParameters
     _cpValidatorScoreParameters = gcpValidatorScoreParameters
+    _cpMaxLockDuration = gcpMaxLockDuration
 
 -- | Convert 'GenesisParametersV2' to genesis data.
 --  This is an auxiliary function since much of the behaviour is shared between protocol versions.